programs/pptx-from-template/pptx-skill-dev/pyproject.toml
[project]
name = "pptx-skill-dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-pptx>=1.0.2",
]
